refactor(livekit): extract ReplaySegmentsService from LivekitReplayService
Moves the segment lifecycle cluster — cleanupOldSegments, cleanupOrphanedSessions
and cleanupRemuxCache crons, segment listing/sorting, segment path resolution and
HLS remux caching — into a dedicated ReplaySegmentsService (507 lines, deps:
egress client, config, database, storage). LivekitReplayService keeps egress
lifecycle + clip capture orchestration and shrinks 1684 -> 1204 lines; the
@Cron decorators relocate with their methods. Corresponding tests moved into
replay-segments.service.spec.ts.

Verified: 125 suites / 2119 tests green, type-check + lint clean, backend boots
with 0 compile errors.
